# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Contacts API",
    "description": "Api for Contacts application",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://cse341-contacts-moso.onrender.com"
    }
  ],
  "paths": {
    "/contacts/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/contactCreate"
              }
            }
          }
        }
      }
    },
    "/contacts/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/contactUpdate"
              }
            }
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "contactUpdate": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "any"
          },
          "lastName": {
            "type": "string",
            "example": "any"
          },
          "email": {
            "type": "string",
            "example": "any"
          },
          "favoriteColor": {
            "type": "string",
            "example": "any"
          },
          "birthday": {
            "type": "string",
            "example": "any"
          }
        }
      },
      "contactCreate": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "any"
          },
          "lastName": {
            "type": "string",
            "example": "any"
          },
          "email": {
            "type": "string",
            "example": "any"
          },
          "favoriteColor": {
            "type": "string",
            "example": "any"
          },
          "birthday": {
            "type": "string",
            "example": "any"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "favoriteColor",
          "birthday"
        ]
      }
    }
  }
}